{"schema":"decodex/commit/1","summary":"Consolidate cohesive Swift app models","authority":"manual"}
